chore : price int 형 통일로 인한 자잘한 수정

+++ src/main/kotlin/com/example/cakezip/service/ShopServiceImpl.kt
@@ -44,7 +44,7 @@ class ShopServiceImpl(
             }
             var info: List<String> = design.split(" ");
             var detail: String = info.get(0)
-            var price: Long = info.get(1).replace("원", "").toLong()
+            var price: Int = info.get(1).replace("원", "").toInt()
             val cakeOptionList: CakeOptionList = CakeOptionList(newShop, OptionTitleType.DESIGN, detail, price);
             cakeOptionListRepository.save(cakeOptionList)
         }
@@ -56,7 +56,7 @@ class ShopServiceImpl(
             }
             var info: List<String> = size.split(" ");
             var detail: String = info.get(0)
-            var price: Long = info.get(1).replace("원", "").toLong()
+            var price: Int = info.get(1).replace("원", "").toInt()
             val cakeOptionList: CakeOptionList = CakeOptionList(newShop, OptionTitleType.SIZE, detail, price);
             cakeOptionListRepository.save(cakeOptionList)
         }
@@ -68,7 +68,7 @@ class ShopServiceImpl(
             }
             var info: List<String> = sheet.split(" ");
             var detail: String = info.get(0)
-            var price: Long = info.get(1).replace("원", "").toLong()
+            var price: Int = info.get(1).replace("원", "").toInt()
             val cakeOptionList: CakeOptionList = CakeOptionList(newShop, OptionTitleType.SFLAVOR, detail, price);
             cakeOptionListRepository.save(cakeOptionList)
         }
@@ -80,7 +80,7 @@ class ShopServiceImpl(
             }
             var info: List<String> = cream.split(" ");
             var detail: String = info.get(0)
-            var price: Long = info.get(1).replace("원", "").toLong()
+            var price: Int = info.get(1).replace("원", "").toInt()
             val cakeOptionList: CakeOptionList = CakeOptionList(newShop, OptionTitleType.CFLAVOR, detail, price);
             cakeOptionListRepository.save(cakeOptionList)
         }
@@ -92,7 +92,7 @@ class ShopServiceImpl(
             }
             var info: List<String> = creamColor.split(" ");
             var detail: String = info.get(0)
-            var price: Long = info.get(1).replace("원", "").toLong()
+            var price: Int = info.get(1).replace("원", "").toInt()
             val cakeOptionList: CakeOptionList = CakeOptionList(newShop, OptionTitleType.CCOLOR, detail, price);
             cakeOptionListRepository.save(cakeOptionList)
         }
@@ -104,7 +104,7 @@ class ShopServiceImpl(
             }
             var info: List<String> = letter.split(" ");
             var detail: String = info.get(0)
-            var price: Long = info.get(1).replace("원", "").toLong()
+            var price: Int = info.get(1).replace("원", "").toInt()
             val cakeOptionList: CakeOptionList = CakeOptionList(newShop, OptionTitleType.LCOLOR, detail, price);
             cakeOptionListRepository.save(cakeOptionList)
         }
